Extract class TaskList
