+++ src/main/java/Storage.java
@@ -45,8 +45,6 @@ public class Storage {
         }
         return "";
     }
-
-
     public static void dumpTask(Task task) {
         try {
             File file_name = new File(DUKE_DUMP_FILE);
@@ -69,4 +67,5 @@ public class Storage {
             e.printStackTrace();
         }
     }
+     public static 
 }
